[웹개발 4주차 미션] 구현 완료했습니다. 리뷰 부탁드립니다.

+++ src/main/java/codesquad/controller/QnaController.java
@@ -2,6 +2,8 @@ package codesquad.controller;
 
 import java.util.NoSuchElementException;
 
+import javax.servlet.http.HttpSession;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -9,9 +11,12 @@ import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.servlet.ModelAndView;
 
 import codesquad.domain.question.Question;
 import codesquad.domain.question.QuestionRepository;
+import codesquad.domain.user.User;
 
 @Controller
 public class QnaController {
@@ -32,40 +37,57 @@ public class QnaController {
 	}
 
 	@PostMapping("/questions")
-	public String ask(Question question) {
+	public String ask(Question question, HttpSession session) {
+		User sessionedUser = (User)session.getAttribute("sessionedUser");
+
+		if (sessionedUser == null) {
+			return "redirect:/login";
+		}
+		question.setWriter(sessionedUser);
 		questionRepository.save(question);
 		return "redirect:/";
 	}
 
 	@GetMapping("/questions/{index}")
-	public String show(@PathVariable Long index, Model model) {
+	public ModelAndView show(@PathVariable Long index, HttpSession session) {
+		ModelAndView mav = new ModelAndView("qna/show");
 		Question question = questionRepository.findById(index).orElseThrow(NoSuchElementException::new);
-		model.addAttribute("question", question);
-		return "qna/show";
+		mav.addObject("question", question);
+		return mav;
 	}
 
 	@GetMapping("/questions/{index}/form")
-	public String updateForm(@PathVariable Long index, Model model) {
+	public String updateForm(@PathVariable Long index, Model model, HttpSession session) {
+		User sessionedUser = (User)session.getAttribute("sessionedUser");
 		Question savedQuestion = questionRepository.findById(index).orElseThrow(NoSuchElementException::new);
-		if(savedQuestion.isSameIndex(index)) {
-			model.addAttribute("question", savedQuestion);
+
+		if (!sessionedUser.isSameId(savedQuestion.getWriter().getUserId())) {
+			return "failed";
 		}
+
+		model.addAttribute("question", savedQuestion);
 		return "qna/updateForm";
 	}
 
-	@PostMapping("/questions/{index}")
-	public String update(@PathVariable Long index, Question updatedQuestion) {
+	@PutMapping("/questions/{index}")
+	public String update(@PathVariable Long index, Question updatedQuestion, HttpSession session) {
+		User sessionedUser = (User)session.getAttribute("sessionedUser");
 		Question question = questionRepository.findById(index).orElseThrow(NoSuchElementException::new);
-		if(updatedQuestion.isSameIndex(question.getIndex())) {
+		if (question.isSameIndex(index)) {
 			question.update(updatedQuestion);
 			questionRepository.save(question);
 		}
 		return "redirect:/";
+
 	}
 
 	@DeleteMapping("/questions/{index}")
-	public String delete(@PathVariable Long index) {
+	public String delete(@PathVariable Long index, HttpSession session) {
+		User sessionedUser = (User) session.getAttribute("sessionedUser");
 		Question question = questionRepository.findById(index).orElseThrow(NoSuchElementException::new);
+		if(!sessionedUser.isSameId(question.getWriter().getUserId())) {
+			return "redirect:/login";
+		}
 		questionRepository.delete(question);
 		return "redirect:/";
 	}
